Create graph links from source and target nodes

This function creates a data frame of graph links from a set of source and target nodes.

Usage

create_graph_links(
  source_nodes,
  target_nodes,
  directed = FALSE,
  unique_links = TRUE
)

Arguments

source_nodes

A vector of source node IDs.

target_nodes

A vector of target node IDs.

directed

If FALSE, the default, the links are undirected. If TRUE, the links are directed.

unique_links

If TRUE, the default, the links are unique. Duplicate links are removed. If FALSE, duplicate links are retained.

Value

A data frame of graph links with columns from and to.

Examples

# Create source and target nodes
source_nodes <- c("jn_000001", "jn_000002", "jn_000002")
target_nodes <- c("jn_000002", "jn_000001", "jn_000003")

# Create graph links (duplicate links are removed)
create_graph_links(source_nodes, target_nodes)
#> Error in create_graph_links(source_nodes, target_nodes): could not find function "create_graph_links"